wip; moved preppers to new package. removing docker dependency for images in local daemon
