+++ pkg/util/tar_utils.go
@@ -68,10 +68,11 @@ func unpackTar(tr *tar.Reader, path string) error {
 
 		// if it's a file create it
 		case tar.TypeReg:
-
+			os.Chmod(target, mode)
 			currFile, err := os.OpenFile(target, os.O_CREATE|os.O_TRUNC|os.O_WRONLY, mode)
 			if err != nil {
 				glog.Errorf("Error opening file %s", target)
+				glog.Errorf("File Mode: %s", mode.String())
 				return err
 			}
 			_, err = io.Copy(currFile, tr)
@@ -106,7 +107,9 @@ func UnTar(filename string, target string) error {
 }
 
 func IsTar(path string) bool {
-	return filepath.Ext(path) == ".tar"
+	return filepath.Ext(path) == ".tar" ||
+		filepath.Ext(path) == ".tar.gz" ||
+		filepath.Ext(path) == ".tgz"
 }
 
 func CheckTar(image string) bool {
